Refactor access management functions and update namespace usage
