+++ bots/wd.php
@@ -10,7 +10,7 @@ include_once __DIR__ . '/../include.php';
 
 use function Publish\GetToken\post_params;
 use function Publish\MdwikiSql\fetch_query;
-use function Publish\Helps\get_access_from_db;
+use function Publish\AccessHelps\get_access_from_db;
 use function Publish\Helps\pub_test_print;
 use function Publish\Helps\get_url_curl;
 
@@ -117,6 +117,7 @@ function getAccessCredentials($user, $access_key, $access_secret)
 {
     if (!$access_key || !$access_secret) {
         $access = get_access_from_db($user);
+        // ---
         if ($access === null) {
             pub_test_print("user = $user");
             pub_test_print("access == null");
