+++ include.php
@@ -11,6 +11,7 @@ include_once __DIR__ . '/bots/mdwiki_sql.php';
 
 include_once __DIR__ . '/config.php';
 include_once __DIR__ . '/helps.php';
+include_once __DIR__ . '/access_helps.php';
 include_once __DIR__ . '/do_edit.php';
 include_once __DIR__ . '/add_to_db.php';
 include_once __DIR__ . '/get_token.php';
