+++ tests/test.php
@@ -5,8 +5,6 @@ ini_set('display_startup_errors', 1);
 error_reporting(E_ALL);
 
 include_once __DIR__ . '/../bots/Citation.php';
-
-// include_once __DIR__ . '/helps.php';
 include_once __DIR__ . '/../bots/md_cat.php';
 include_once __DIR__ . '/../bots/text_fix_refs.php';
 include_once __DIR__ . '/../bots/remove_duplicate_refs.php';
